Move roles and permissions validation to a mixin
