+++ lib/auth0/api/v2/roles.rb
@@ -1,8 +1,12 @@
+require 'auth0/mixins/validation'
+
 module Auth0
   module Api
     module V2
       # Methods to use the users endpoints
       module Roles
+        include Auth0::Mixins::Validation
+
         attr_reader :roles_path
 
         private
